stick invariant <'brand> tag onto Scope and CoreExt
When we update to rust-simplicity 0.6, we will need to stick an invariant
lifetime (conventionally called 'brand) onto a bunch of types to ensure
that constructed nodes from different type inference contexts are not
mixed.

For the types we own, do this ahead of the actual update, to separate out
the noise from the "real" changes.