named: replace ad-hoc ConstructNode/CommitNode/WitnessNode with generic ones
Now we have a marker type WithNames which wraps a marker type from
rust-simplicity, erasing its disconnect nodes and replacing its witness
nodes with names. This gets us an analogue of ConstructNode, CommitNode,
WitnessNode, RedeemNode, etc., all at once.

Then replace:

* to_commit_node, whose role was to forget names and finalize types, with
  a general method forget_names, which just forgets the names (then the
  user can call .finalize_types() themselves if they want)
* to_witness_node, whose role was to replace names with actual witnesses,
  with a general method populate_witnesses, which does this

(Actually, `to_witness_node` is not very general at all. It specifically
converts a WtihNames<ConstructNode> to an unnamed ConstructNode. In this
commit, it's possible to implement it in a general way, converting an
arbitrary Node<WithNames<M>> to a Node<M>. But in a couple commits, when
we replace ConstructNode with CommitNode in CompiledProgram, we will be
forced to fix its signature to convert a CommitNode to a RedeemNode. The
diff will be easier to read starting from this non-generic version than
it would be if we started from the generic version.)

Along the way, use core::convert::Infallible for error types so that we
don't need to unwrap errors.

This leaves the precondition on `populate_witnesses` that the user call
`is_consistent` on witness nodes before converting from SimplicityHL to
Simplicity. I believe I can remove this precondition and return an error
if the witness values don't match the Simplicity node, but I need to
update the library to get the IncompleteType type first, so that I can
return a useful error. So I will defer this improvement.