replace Construct with Commit in CompiledProgram
This commit represents an important change: in the user-facing types we
are no longer holding ConstructNode objects, which mean that we are not
handing the user incomplete types which are tied to a specific inference
context.

This means that within our code, whenever we need to manipulate types, we
need to create a new local type inference context which is then dropped
at the end of the function. We'll need this when we update rust-simplicity
to a version that has only locally-scoped context objects.

The bulk of the diff is in src/named.rs, changing the populate_witness
method to convert a CommitNode to a RedeemNode, which is mostly just fixing
type signatures by chasing compiler errors. Sorry for the noise. I also
add a new stringly-typed error, which I'm not thrilled with, but this
crate is already full of them.