move builtins from crate root to private module in compile
This module was introduced in #145 as a public module. But actually it is
only ever used by the compiler, and it doesn't really make sense outside
the compile because of its use of `ProgNode`.

We may want to later make the builtins module be an extension trait of
Constructible or something, but as long as it depends on ProgNode it
should not be public (and it cannot be public, after the next commit).