array_fold: handle powers of two correctly
Our array_fold function explicitly computes compilations for trees of
size 2^n, but only uses this to shortcut computations for the special
case n == 1. This leads to panics wherein we wind up trying to "fold"
a tree of size 0.

Shortcut all powers of 2 so that we don't ever hit this case.