add tree-sitter as main parsing option
Add the tree-sitter grammar parser and use it for the main LSP functions.

Functions changed:
- completion: completion now also accepts visible variables for
  completion, which we obtain with tree-sitter.
- hover: now it searches inside the tree-sitter AST instead of the AST
  generated by the SimplicityHL compiler. Also added more hover options,
  such as types and variables.
- goto-definition and references: also updated to search using the tree-sitter
  AST and to use tree-sitter’s query features for finding definitions and
  references.

As we now use the AST provided by tree-sitter, the functions working with
the compiler AST are removed, as is the `miniscript` crate.

Added the `documentation` module, because all SimplicityHL function
documentation was previously inside the `completion` module, which was
confusing since this documentation is also used by other functions,
such as `hover`.